[libs] E: Move Binding for connect()
This commits moves the binding for the connect() system calls to the
syscall crate
