+++ src/libs/syscall/src/sys/socket/bindings/mod.rs
@@ -7,6 +7,7 @@
 
 pub mod accept;
 pub mod bind;
+pub mod connect;
 pub mod listen;
 pub mod setsockopt;
 pub mod socket;
